Custom kernel configuration
Added kernel configuration (defconfig) in linux-raspberrypi_4.19.bbappend.
This will issue an warning during build:
  defconfig detected in WORKDIR. bcm2709_defconfig skipped
Warning is generated by kernel-yocto.bbclass, but can be ignored.
The actual presence of this warning tells us that we are using the new defconfig.

The defconfig in this commit make some simple changes to the kernel configuration.
Some  drivers are changed to be included in the kernel instead of being modules.
Also removed support for IPv6 and some other stuff.